shifted profile and destroy methods from home to auth/user controller
