+++ app/Http/Controllers/Auth/LoginController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Auth;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
+use Illuminate\Support\Facades\Auth;
 
 class LoginController extends Controller
 {
@@ -37,4 +38,19 @@ class LoginController extends Controller
     {
         $this->middleware('guest')->except('logout');
     }
+
+    /**
+     * It will handle the autehnticate request.
+     */
+
+     public function authenticate(Request $request)
+     {
+         $credentials = $request->only('email', 'password');
+
+         dd($credentials);
+         if (Auth::attempt($credentials)) {
+
+            return redirect()->intended('/home');
+         }
+     }
 }
